添加切换主题色过渡
